+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.7.0
 
 * :arrow_up: Updating to Kotlin 1.2.71
+* :fire: Formatx totalTimeZH and totalTimeZHSHort method default parameter '分' to '分钟'
 
 
 ## v0.6.9